requirements.txt
python-jobspy>=1.1.80
feedparser>=6.0.11
supabase>=2.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
